requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-google-genai
langgraph
langchain-community
faiss-cpu
PyPDF2
python-multipart
pydantic
sqlalchemy
python-dotenv
requests
chromadb